+++ src/main/resources/db/migration/V00003__delete_article_previews_table.sql
@@ -1,0 +1,3 @@
+ALTER TABLE blog_sub DROP FOREIGN KEY blog_sub_ibfk_1;
+
+DROP TABLE IF EXISTS article_previews;
