--- src/test/resources/sql/article_previews.sql
@@ -1,10 +1,0 @@
-id INT,
-                           main_header VARCHAR(100),
-                           sub_header VARCHAR(100),
-                           text TEXT,
-                           link VARCHAR(200),
-
-INSERT INTO article_previews VALUES
-                          (1, 'main header 1', 'sub header 1', 'text 1', 'https://link1.com'),
-                          (2, 'main header 2', 'sub header 2', 'text 2', 'https://link2.com'),
-                          (3, 'main header 3', 'sub header 3', 'text 3', 'https://link3.com'),
